pyproject.toml
[project]
name = "class-hw4-04-11-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.41.3",
    "flask>=3.1.2",
    "python-dotenv>=1.0.0",
]
